+++ sdk/include/checkup_weather/utility.h
@@ -19,6 +19,6 @@ namespace checkup
     const char* category_to_string(TWeatherCategory category);
     bento::Vector3 temperature_to_color(float temperature);
     bool is_day(uint32_t time, uint32_t sunrise, uint32_t sunset);
-    void time_to_string(uint32_t time, bento::DynamicString& outputString);
+    void time_to_string(int64_t time, bento::DynamicString& outputString);
     void duration_to_string(uint32_t duration, bento::DynamicString& outputString);
 }
